requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
line-bot-sdk==3.8.1
pytest==8.2.0
